Refactor drawing logic into Drawer
