+++ src/visualizers/animations/empty-animation.ts
@@ -1,0 +1,10 @@
+import Animation from "visualizers/animations";
+
+export default class EmptyAnimation implements Animation {
+    run(dt) {
+    }
+
+    isFinished() {
+        return true;
+    }
+}
