+++ src/visualizers/drawables/numbered-circle.ts
