+++ src/visualizers/drawers/index.ts
@@ -1,0 +1,11 @@
+import Step from 'algorithms/step';
+import Animation from 'visualizers/animations'
+
+export default interface Drawer {
+    handleStep(step: Step): Animation;
+    terminate(): void;
+}
+
+export interface DrawerConstructor {
+    new (width: number, height: number): Drawer;
+}
